Move parseSources.ts
